requirements.txt
opencv-python
ultralytics
pyyaml
mypy
torch
torchvision
torchaudio
fastapi
sse_starlette
imutils
numpy
pygments
